Mise à jour de la dépendance zustand vers la version 4.5.6, refactorisation du store d'authentification et ajout des types TypeScript pour les réponses d'authentification.
