+++ front/package.json
@@ -16,7 +16,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.22.2",
-    "zustand": "^4.5.1"
+    "zustand": "^4.5.6"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.1",
