--- front/src/store/authStore.ts
@@ -1,38 +1,0 @@
-import { create } from 'zustand';
-import { persist } from 'zustand/middleware';
-
-interface User {
-  id: string;
-  email: string;
-}
-
-interface AuthState {
-  user: User | null;
-  isAuthenticated: boolean;
-  login: (email: string, password: string) => Promise<void>;
-  logout: () => void;
-}
-
-export const useAuthStore = create<AuthState>()(
-  persist(
-    (set) => ({
-      user: null,
-      isAuthenticated: false,
-      login: async (email: string, password: string) => {
-        // TODO: Replace with actual API call
-        if (email && password) {
-          set({
-            user: { id: '1', email },
-            isAuthenticated: true,
-          });
-        }
-      },
-      logout: () => {
-        set({ user: null, isAuthenticated: false });
-      },
-    }),
-    {
-      name: 'auth-storage',
-    }
-  )
-);
